# openapi.yaml
openapi: 3.0.0
info:
  title: VolunChain API
  description: |
    API for managing projects in VolunChain with versioning support.

    ## API Versioning
    This API uses URL path versioning with the following structure:
    - `/api/v1/` - Current stable API version
    - `/api/v2/` - Reserved for future expansion

    All endpoints are namespaced under their respective version paths.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server (v1)
  - url: http://localhost:3000/api
    description: API root (version info)
paths:
...
  /auth/protected:
    get:
      summary: "Access a protected route"
      operationId: "protectedRoute"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Protected resource accessed successfully"
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
...